try to prevent github node reboots

+++ internal/reboot/rebooter.go
@@ -18,6 +18,21 @@ type Rebooter interface {
 }
 
 var _ Rebooter = &watchdogRebooter{}
+var _ Rebooter = &disabledRebooter{}
+
+type disabledRebooter struct {
+	err error
+}
+
+// NewDisabledRebooter returns a rebooter that fails closed without attempting
+// either a watchdog or software reboot.
+func NewDisabledRebooter(err error) Rebooter {
+	return &disabledRebooter{err: err}
+}
+
+func (r *disabledRebooter) Reboot() error {
+	return r.err
+}
 
 // watchdogRebooter uses a watchdog for triggering reboots
 type watchdogRebooter struct {
@@ -39,10 +54,10 @@ func (r *watchdogRebooter) Reboot() error {
 	if r.wd == nil {
 		r.log.Info("no watchdog is present on this host, trying software reboot")
 		//we couldn't init a watchdog so far but requested to be rebooted. we issue a software reboot
-		return r.softwareRebootHook()
+		return r.trySoftwareReboot()
 	} else if r.wd.Status() == watchdog.Malfunction {
 		r.log.Info("watchdog is malfunctioning on this host, trying software reboot")
-		return r.softwareRebootHook()
+		return r.trySoftwareReboot()
 	}
 
 	//Watch dog is rebooting, wait to make sure watchdog is rebooting properly otherwise intervene with software reboot
@@ -50,12 +65,12 @@ func (r *watchdogRebooter) Reboot() error {
 	case watchdog.Triggered:
 		r.log.Info("watchdog is triggered, waiting for watchdog reboot to commence")
 		if r.isWatchdogRebootStuck() {
-			return r.softwareRebootHook()
+			return r.trySoftwareReboot()
 		}
 		return nil
 	case watchdog.Disarmed:
 		r.log.Info("watchdog failed to start, trying software reboot")
-		return r.softwareRebootHook()
+		return r.trySoftwareReboot()
 	case watchdog.Armed:
 		// we stop feeding the watchdog for a reboot
 		r.wd.Stop()
@@ -68,15 +83,34 @@ func (r *watchdogRebooter) Reboot() error {
 	}
 }
 
-// softwareReboot performs software reboot by running systemctl reboot
+func (r *watchdogRebooter) trySoftwareReboot() error {
+	softwareRebootEnabled, err := watchdog.IsSoftwareRebootEnabled()
+	if err != nil {
+		return err
+	}
+	if !softwareRebootEnabled {
+		return errors.New("software reboot is disabled")
+	}
+	return r.softwareRebootHook()
+}
+
+// softwareReboot performs a software reboot through the host's sysrq interface.
 func (r *watchdogRebooter) softwareReboot() error {
+	softwareRebootEnabled, err := watchdog.IsSoftwareRebootEnabled()
+	if err != nil {
+		return err
+	}
+	if !softwareRebootEnabled {
+		return errors.New("software reboot is disabled")
+	}
+
 	r.log.Info("about to try software reboot")
 	// privileged:true required to run this
 	rebootCmd := exec.Command("/usr/bin/nsenter", "-m/proc/1/ns/mnt", "/bin/bash", "-c", "echo b > /proc/sysrq-trigger")
 
 	if err := rebootCmd.Run(); err != nil {
 		r.log.Error(err, "failed to run reboot command")
-		// TODO retry because of this?
+		return err
 	}
 	return nil
 }
